Simplify icon typings
