+++ packages/gamut-icons/icon-template.js
@@ -11,26 +11,23 @@ function iconTemplate(api, opts, { jsx /* imports, props, exports */ }) {
     import * as React from 'react';
     import { ClassNames } from '@emotion/react';
     import { GamutIconProps } from '../../types';
-    import { getAttrValue, getForwardableProps, IconStyleProps, iconStyles } from '../../props';
-    const LocalIcon = React.forwardRef<SVGSVGElement, GamutIconProps>(({
+    import { getForwardableProps, GamutIconProps, iconStyles } from '../../props';
+
+    export const ${exportName} = React.forwardRef<SVGSVGElement, GamutIconProps>(({
       title = "${title}",
       titleId,
       size,
-      color,
-      width,
-      height,
       className: cn,
       ...rest
     },
       svgRef
     ) => {
-      const props = getForwardableProps(rest) as Omit<GamutIconProps, keyof IconStyleProps>;
+      const props = getForwardableProps(rest);
       return <ClassNames>{({ css, cx, theme }) => {
-        const classNames = cx(cn, css(iconStyles({ ...rest, width, height, theme })));
+        const classNames = cx(cn, css(iconStyles({ ...rest, theme })));
         return ${jsx}
       }}</ClassNames>;
     });
-    export const ${exportName} = LocalIcon;
   `;
 }
 module.exports = iconTemplate;
