+++ packages/gamut-icons/src/props.ts
@@ -2,7 +2,26 @@ import { styledOptions, system } from '@codecademy/gamut-styles';
 import { StyleProps, variance } from '@codecademy/variance';
 import { pickBy } from 'lodash';
 
-import { GamutIconProps } from './types';
+export interface IconStyleProps extends StyleProps<typeof iconProps> {}
+
+export interface GamutIconProps
+  extends Omit<React.SVGProps<SVGSVGElement>, keyof IconStyleProps>,
+    IconStyleProps {
+  /**
+   * A suffix added to the end of the unique generated ID for the icon. This is useful if you have multiple of the same icon on the page and need to pass accessibility guidelines.
+   */
+  titleId?: string;
+  title?: string;
+  ref?: React.Ref<SVGSVGElement>;
+  /**
+   * See:  https://github.com/DefinitelyTyped/DefinitelyTyped/pull/42404
+   * Generated definitions may not match with other vesions @types/react in ^16.9 this ensures that the type exists and is optional as it would in >16.9.21
+   */
+  path?: string;
+  size?: number | string;
+}
+
+export type ForwardableProps = Omit<GamutIconProps, keyof IconStyleProps>;
 
 export const iconProps = variance.compose(
   system.layout,
@@ -11,19 +30,7 @@ export const iconProps = variance.compose(
   system.positioning
 );
 
-export interface IconStyleProps extends StyleProps<typeof iconProps> {}
-
 export const iconStyles = (props: IconStyleProps) => iconProps(props);
 
-export const getForwardableProps = (props: GamutIconProps) =>
+export const getForwardableProps = (props: GamutIconProps): ForwardableProps =>
   pickBy(props, (value, key) => styledOptions.shouldForwardProp(key));
-
-export const getAttrValue = (prop: IconStyleProps['width' | 'height']) => {
-  switch (typeof prop) {
-    case 'string':
-    case 'number':
-      return prop;
-    default:
-      return undefined;
-  }
-};
