+++ packages/gamut-icons/svgr.config.js
@@ -7,19 +7,13 @@ module.exports = {
   ref: true,
   svgProps: {
     className: '{classNames}',
-    width: '{size || getAttrValue(width) || "16px"}',
-    height: '{size || getAttrValue(height) || "16px"}',
+    width: '{size || "16px"}',
+    height: '{size || "16px"}',
     fill: 'currentColor',
     role: 'img',
   },
   prettierConfig: {
     parser: 'typescript',
-  },
-  replaceAttrValues: {
-    currentColor: 'currentColor',
-    '#000': 'currentColor',
-    '#111': 'currentColor',
-    '#444': 'currentColor',
   },
   template,
   indexTemplate,
